requirements.txt
tensorflow-gpu==2.1.0
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
pyyaml==3.12